Cloud overlay resets shader's offset parameter on save
Previously, since the cloud overlay was running in the inspector, the offset of the clouds would change every frame. That's great to see how it looks, but since that's a value that is persisted in the scene, it was causing changes in the frays_end scene file.

Now, the clouds will still move across the level as usual, but on save the offset parameter of the shader gets reset to 0 so it's always the same value and we don't commit new unwanted changes to fray's end by mistake.