SceneSwitcher: Start async scene load during transition
Since the fade-out transition between scenes lasts 1 second, it is worthwhile to
start loading the next scene before beginning the transition, and only wait for
loading to finish once the fade-out has finished. For smaller scenes, 1 second
is long enough for the scene to load; for Fray's End, it's a significant
fraction of the total loading time.

Helps https://github.com/endlessm/threadbare/issues/368