Projectiles: ignore collisions with elevated terrain until projectile escaped that terrain
Collisions with walls are now handled by raycast instead of by the rigidbody itself.

On spawn, the projectile detects which tiles of each tileset is 'touching', and if any of those tiles is an elevated terrain, it adds that tilemap layer to a collision exception list.

As soon as the projectile isn't touching any tile from a tilemap in the collision exception list, that exception gets removed.

Had tocheck positions instead of using the raycast itself on spawn because not all tiles with elevated terrain have collisions: for example, the projectile could have spawned in the middle of a wide mountain and it'd be in elevated terrain but it wouldn't find any collisions until it got to the edge of the cliffs.

Also, used several points to find tiles in its bounding box because if I only checked the origin position of the projectile, it could spawn just at the side of a cliff, not add it to the exception list, and then be impulsed into it, causing the projectile to get stuck in the cliff. With this implementation, it is a bit more generous with what tilemap layers will end up in the collision exception list.

Related to https://github.com/endlessm/threadbare/issues/277