StoryQuestStarter: Disable further interactions when starting quest
When the player has chosen a dialogue path that leads to a quest, there is no
need to reshow the interaction prompt at the end of the dialogue: the scene will
change at that point.

Currently, this is achieved by checking whether player input is paused before
showing the prompt, but this special case will be removed in a future commit.
Instead, disable the InteractArea entirely in that case.