Ink Combat: Redesign Round 3
In order to make this bug impossible, changed the design of the level so
ink drinkers that are on top of the stone walls don't have to shoot
"backwards".

Since I was changing the level around, also placed the ink wells further
apart from each other and placed one of them in such a way that the
player will probably need to bounce the ink blobs against the walls to
hit it.