Checkpoint: Add interaction
This serves as a way to explain to the player what the checkpoint does, without
completely breaking the forth wall.

While trying to think of what the right verb might be, one of my favourite
moments from Day of the Tentacle came to mind, and I couldn't resist adding an
easter egg.