Adjust first checkpoint to be out of sight of guards
Previously, the first checkpoint in the stealth level was position such that the
second guard would see if you if you respawn there and don't move.

This seems cruel, particularly now that the player can interact with the
checkpoint and potentially be mid-conversation when they are caught.

Move the checkpoint west a bit.