Make background music continue between levels
Add a new global, MusicPlayer, that holds an AudioStreamPlayer that is accessible from everywhere.

Change the code in BackgroundMusic so it uses that AudioStreamPlayer instead of an internal one. Also, change the code so it checks if the AudioStream it will play is the same as the one that is currently playing, if that's the case, do nothing.

When transitioning out of a level, the BackgroundMusic node starts a tween that mutes the global AudioStreamPlayer by tweening its volume_db to -100 (linear_to_db(0.00001)). On _ready() of the next BackgroundMusic node, the AudioStreamPlayer volume_db would tween back to 0.0. This solves 2 things:
- We have fade out and fade in between scenes.
- If a level doesn't have background music, we don't have to do anything
  special in that level because the audio stream will be muted because of the
  volume_db. The only gotcha with this is that if we go from level A
  with audio stream 1, to level B with no music, to level C with audio
  stream 1 again, the audio stream will resume where it was playing
  instead of starting again.

Fixes https://github.com/endlessm/threadbare/issues/335