Add volume controls to pause overlay
In order to be able to use the same controls as the title screen, I redesigned
the pause overlay a little. It now uses the same theme (the very misleadingly
named dialogue theme...) and a semitransparent version of the fabric texture
that is used for the title screen background.

Fixes https://github.com/endlessm/threadbare/issues/8
Fixes https://github.com/endlessm/threadbare/issues/336