Fix trees sway in sync
In the trees shaders we use the world position as a parameter to sample a noise to get the wind parameters. There was an issue though: Since the textures get sampled using UVs, we need to convert the world coordinates into UV coordinates.

The use of the `fract` function was ok, but we should divide the world position by a "scale" which will decide how big or small is the projection of the texture on the world. That scale will decide how "uniform" is the movement of the trees, large values being more uniform, and small values less.

There's already a parameter for that called "phase_period" but it wasn't being used.

This commit makes use of that parameter fixing the "sway in sync" behaviur of the trees. Also, the noise of the wind of trees is tuned so the difference in wind is softer.

Fixes #307