Fill game: Remove ink_combat.gd script
This script was only re-emitting the goal_reached signal that was
already in the FillGameLogic node. And was presenting an initial
dialogue that can be moved to the FillGameLogic node.

Move the initial dialogue of combat round 1 to the FillGameLogic node.

Adjust the combat rounds to connect the opening of the exit door (round
1 and 2) or the reveal of the collectible (round 3) to the goal_reached
of the FillGameLogic node.

Also, enable autostart in enemies of combat 2 and 3, because there is no
dialogue on them.