Stealth game: Join logic in a single script
And move it from the Lore quest folder to the scenes/game_logic folder.

The logic was split in 2 scripts. Merge them so we have a single node
with the logic.

Also use the naming convention of starting functions with an underscore
in the private methods of the script.

Add a class name to the script and add it as a node to the Stealth
scenes.