Add addon to change textures from SpriteFrames
The addons affects any SpriteFrames resource by exposing some 'virtual' properties (one for each texture used in the SpriteFrames) in the inspector, and if any of those texture is changed, it updates every animation in the SpriteFrames that used accordingly.

The way to access those properties is either by expanding the SpriteFrames that are exposed by any object, or by opening a SpriteFrames resource directly in godot.

Added some safeguards:

- only lets users swap a texture by another texture of the same size.

- doesn't let the user swap a texture by an already used texture (this messed up undoing the operation).

- doesn't let the user clear the texture.

Fixes #173