Filling Barrel: determine needed fills from animation
Previously, the amount of times a filling barrel needed to be filled was hardcoded to 3, and the animation for it could be configured, but it would always need to have 4 frames.

Now, that requirement was relaxed, the FillingBarrel exposes a SpriteFrames instead of a texture, and the amount of times it needs to be filled can be configured.