Switch all res:// preload references to UID paths
This makes the scripts immune to file renames.

It is also less readable, but the Godot editor understands these paths
and you can Ctrl-click on them in the script editor to see the resource
open in the Inspector. From there you can also use the top bar menu to
show the file in the Filesystem dock. Or, if it is a scene, it opens in
a new tab directly, then right-clicking in the root node it can be shown
in the Filesystem dock.