Link Sokoban puzzles to the Eternal Loom
When the player interacts with the Eternal Loom while in possession of the
threads of Memory, Imagination, and Spirit (after completing a quest) we
previously played an animation of the threads entering the loom, and then some
dialogue spoken by the Elders.

After the animation plays, switch scenes to a series of three Sokoban puzzles
where the player repairs the loom using the threads they have collected. Once
they have completed all three puzzles, return to Fray's End and play the Elder
dialogue.

Fixes https://github.com/endlessm/threadbare/issues/440