Add flip & offset properties as exposed properties in decoration.gd
In order to allow for more variation, decorations can have their flip and offset properties so they can be tuned by level editors.

The flip_h and flip_v properties allow to reuse the same decoration, but adding a bit more variation. The offset, on the other hand, allows to stack decorations without messing with the y-sort.