Eliminate MusicalRockXylophone
This was a wrapper around a fixed layout of 6 rocks, and served only to:

- Discover its rock children
- Proxy play() calls through to the correct rock
- Proxy note_played signals outwards

The fixed set of rocks was a barrier to varying the puzzle in different
levels.

Inline the child rocks into the two scenes that used the
musical_rock_xylophone scene. Move the rock discovery, playing, and
listening out to MusicPuzzle.

To allow the rocks to be grandchildren of MusicPuzzle rather than
direct children (so they can be grouped together into an intermediate
node if desired), introduce a new sequence_object group, place the rock
scene into it, and use this when discovering rocks. The name hints at a
future where the rocks may not be rocks, and the sequence may not
necessarily be a musical melody.

Fixes https://github.com/endlessm/threadbare/issues/519