BonfireSign: Remove reference to MusicPuzzle
Previously, there was a circular chain of references necessary for signs
to work:

- MusicPuzzle has a reference to a list of SequencePuzzleSteps
- SequencePuzzleStep has a reference to a BonfireSign
- BonfireSign had a reference to MusicPuzzle

The last reference was used to trigger a demo of the sign's sequence.

I found this cyclic reference ugly; it's also one more thing that can be
misconfigured when creating a sequence puzzle.

Instead, have BonfireSign emit a signal when the sequence should be
demoed. Connect to this signal in the MusicPuzzle logic, which avoids
having to later match the sign up to its puzzle step.

It's a little awkward because we want to delay the end of the
interaction until the demo is complete, but the demo handler is an async
function, and Godot does not have a way for a signal emitter to await
all connected handlers completing. So instead, make BonfireSign wait for
a demonstration_finished() method being called, but only in the case
where a handler is connected to the signal.