Shaker: Fix runtime type warnings
There were many warnings about functions and variables without explicit
types. Buried in there was a warning about the new_rotation variable
being unused. In fact this turned out not to be a bug – at the point
where it should be used, the same value was recomputed instead – but
cleaning up the warnings will make it easier to spot actual problems in
new code in future.