Fill game: Fix FillingBarrel hitbox collision layer
The hitbox area shouldn't be in the walls layer. Only the capsule shape
of the static body should be a wall.

The hitbox area should only mask the projectiles layer because the
projectile checks if it collided with FillingBarrel in the handler of
the "body_entered" signal.

This regressed in f8e5eed8 while changing the collisions for the capsule
shape.