Fill game: Allow modding the projectile trail
Like for the other visual FXs in the projectile. Allow passing a scene
for changing the trail. Or to clear the property for removing the
default trail.

Move the existing particles to a separate scene under game_elements/fx/
. Add a marker in its place for using it as reference when
instantiating.

Export the same property in the ThrowingEnemy, because projectiles are
spawned there.