Convert river sound to Ogg Vorbis
Previously the sound of the river in the LoreQuest music puzzle was a 17
MB WAV file. By way of comparison, the project source tree is 57 MB in
total, so this is 30% of the size of the project. (It was converted at
export time to QOA, but in the context of hundreds of learners cloning
the source tree, 17 MB per clone adds up fast.)

Convert it to Ogg Vorbis with the default quality setting (3):

    oggenc \
        -a 'Pfannkuchn' \
        -t 'Small River 1 - Fast - Semi Close'  \
        ./assets/third_party/sounds/ambient/459405__pfannkuchn__small-river-1-fast-semi-close.wav

resulting in a more reasonable 750 KB file.

https://github.com/endlessm/threadbare/issues/568