HintSign: Make demonstrating sequence optional
Previously, interacting with a hint sign would always demonstrate the
sequence. Now that it is possible to supply a `hint` animation to play
on the sign itself (as is done in the template), some level designers
may prefer to not demonstrate the sequence and make the puzzle more than
a memory/imitation game.

For example, the template puzzle has two signs. Each sign shows a
different series of symbols that have a natural correspondence to a
coloured object (e.g. green = leaf). You could imagine changing this
puzzle so that the player learns the correspondence from the first sign
(which demonstrates the sequence of objects) and then has to interpret
the symbols alone on the second sign. (They can replay the first sign if
they have forgotten.)

Fixes https://github.com/endlessm/threadbare/issues/525

Co-authored-by: Manuel Quiñones <manuq@endlessos.org>