Player: Add dust particles when running
Add a small spritesheet with a dust cloud. Use it to add a new
GPUParticles2D node to the player scene.

Add a script to the particles, so they don't emit when the player is not
running. And use the player velocity to set the direction of the
particles.

Make the particles material semi transparent, and with an "Add" blend
mode.

Add a bit of randomness to the particle effect. In scale and in initial
speed.

Fixes https://github.com/endlessm/threadbare/issues/245