StoryWeaver: Use different animation for attack
Use the animation named "Attack 02" in the Aseprite source file for the
SpriteFrames animation "Attack 01".

Ideally we would change the AnimationPlayer in the Player scene to just
use Attack 02. But we can't do that without also adding back "Attack 02"
to the template for the player. That "Attack 02" animation was removed
from the player template upon request, to avoid confusion in learners,
as the animation wasn't used.

So, swap animations "Attack_01" and "Attack_02" in StoryWeaver
SpriteFrames, in all color variations.

This is confusing, but may be the only thing we can do at this point.