FixedSizeLabel: Resize container to label
This is a workaround for an upstream bug:
https://github.com/godotengine/godot/issues/100626

Connect the label size changed signal to a handler, that calls
reset_size() in the PanelContainer

Also, change the minimum size of the PanelContainer to the current size,
and shrink center both the MarginContainer and the Label nodes. The
offsets in the diff have changed automatically because of this.

Fixes https://github.com/endlessm/threadbare/issues/608