ci: Move Flatpak build to main export workflow
Previously, this workflow was triggered by a successful run of the
export workflow on the `main` branch. But this means that it was not
triggered for releases.

Instead, merge the Flatpak build workflow with the main export workflow.
Make the Flatpak job dependent on the build job, with the same
conditions as before, namely only running on the upstream repo, and only
for builds on main or releases.