Add plugin to automate the creation of StoryQuests
And activate it in this project.

The plugin adds a Tool menu option in the Editor that opens a pop-up
dialog. The dialog ask the user for a StoryQuest title and description,
and derives the StoryQuest folder path from the title.

After accepting the dialog, creates a StoryQuest folder with a
quest.tres file, and 2 scenes: intro and outro. Each scene is duplicated
from the template StoryQuest, along with external resources: one
dialogue for each, and an image texture for the intro.

Please note that the Player spriteframes in both intro and outro is are
still referencing the StoryQuest template.

Also note that the dialogues are copied from the template, so for
instance the intro one says "Welcome to the StoryQuest template" even
when that's not the case.

Resolves https://github.com/endlessm/threadbare/issues/747