EnemyDestroysFloor: Allow enemy to walk across the void
Previously the navigation region was (roughly) the same shape as the
player's walkable area.

I experimented with adding navigation data to the TileSet but in this
case that's not desirable. The enemy would destroy its own navigation
region and quickly run to the nearest island of navigable terrain (often
away from the player). The navigation area should persist even after the
enemy has destroyed the floor.

In fact: if the enemy can float across the void, they should not be tied
to the walkable area at all. Adjust the navigation area to cover the
whole void (give or take). Reduce the enemy's speed, but have them
retarget the player more frequently; and make the path more winding.
This means that the enemy can now cut corners that the player cannot,
making it a little harder to outrun the enemy even though the player is
faster.