Tileset: Add tears to terrains, with low probability
For each of the grass, sand and dirt terrains, add the 3 tiles with
tears with the same peering bits as the center tiles. Set the
probability of the smaller tears to 0.02, and the probability of the
larger tear to 0.005. These numbers are based on wholly unscientific
experimentation of what looks good when I draw big areas.

The light grass tiles do not have versions with tears so that terrain is
not updated.