Revert changes to base guard
The template guard was inadvertently modified, rather than modifying a
copy of it. Then later the copy was modified. But somehow the duplicate
was the one that kept the original uid://ovu5wqo15s5g that is referenced
in guard.gd. Swap the UIDs around.

Similarly the guard scene was mistakenly updated to reference the
equipo2 guard.