Remove duplicate of storyweaver_blue SpriteFrames
One scene in this quest has a custom version of StoryWeaver (carrying
Apolo on their back) but the SpriteFrames used by the other 4 scenes is
identical to storyweaver_blue.tres.

Replace those four references and delete the copy. This will reduce the
work necessary as and when we adjust StoryWeaver in future, e.g.
renaming attack & repel animations.