Add character walk behaviors
These nodes are pure logic, intended to provide an easy way of creating
characters that move around the world with pure setup and no scripting. A more
advanced use is to have a script that can switch between these simple behaviors
to create more complex behaviors.

They are:
- BounceWalkBehavior: Make the character move and bounce on walls.
- ErraticWalkBehavior: Make the character walk around erratically.
- FollowWalkBehavior: Make the character follow a target.
- PathWalkBehavior: Make the character walk along a path.
- InputWalkBehavior: Control the character with input actions to walk and run.

They only take care of controlling the movement of the parent CharacterBody2D
node by setting the velocity and calling move_and_slide. For aesthetics, add a
CharacterSpriteBehavior. This is based in current player_sprite.gd and improved.
In the future, we may also want to support a behavior for animations done with
AnimationPlayer because they can include sound effects that match the visuals.

The walk behavior nodes have exported properties to configure them. There is
quite some repeated code in those. We may consider a single shared resourse
exported.

The walk behavior nodes try to emit as many signals as possible for the
different situations, like when the character turns around when walking on a
path or when the character gets stuck while moving. A script could connect to
these signals to make more interesting behaviors.

The "is stuck" logic is inspired by the one in guard_movement.gd, but changed
to use:
- The boolean returned by move_and_slide() that tells if a collision happened.
- The CharacterBody2D.get_real_velocity()
- The CharacterBody2D.is_on_wall()

For the latter, the character must be properly configured for a top-down game
with motion_mode set to MotionMode.MOTION_MODE_FLOATING.

Resolves https://github.com/endlessm/threadbare/issues/1105