Add new Void Runner level
This is a prototype of a new mechanic, where the player has to outrun an
expanding void in order to get from A to B. If the vortex that is
spreading the void engulfs the player, they unravel; as the vortex
travels across the map, it spreads the void, hiding any props beneath.
It's somewhat inspired by the levels in Crash Bandicoot where a rock is
rolling towards the player.

The vortex moves at the same speed as the player, but speeds up to the
player's running speed if they get too far ahead. Because the player can
run continuously, without consuming stamina, the vortex has to move at
the running speed when the player is running; but I wanted to make it
possible to beat the level without running. It is possible, though it's
a bit easier if you hold down the run button throughout.

Most of the scene starts off covered up by the void. When the player
reaches their goal and interacts with an NPC, the void is pushed back,
revealing new areas of the environment.

This is currently the only challenge of a new LoreQuest. Since all
quests must grant the player three threads, interacting with the monk
grants the player 2 of the 3 threads. The intention is that we'll add
two more challenges to this quest: the second introducing another new
mechanic, and the third combining the two.

All art, dialogue, animations are placeholders, though I couldn't resist
dressing the scene quite thoroughly.

The quest is named and described without using the letter ‘e’ in homage
to La Disparition by Georges Perec, translated into English as A Void,
which I have never read.