SceneSwitcher: Respond to manual hash changes
Previously, if you manually edited the URL hash in the browser address
bar and pressed Enter, nothing would happen. This made it harder to
manually construct the URL for a scene: you received no feedback as to
whether you constructed the URL correctly.

We can monitor the window's hashchange event to be notified when the URL
hash changes. However, this event is also emitted when the hash is
changed by this code, so we need to take care not to enter an infinite
loop in that case.

Store the URL that we most recently set. If we observe the URL changing
to something different, then this must be user-initiated; switch to that
scene.