Speed up typing out dialogue
By default, dialogue is shown at 0.02 seconds per "step" (which I think means
character), with a 0.3 second pause each time a full stop, question mark or
exclamation mark is encountered. Looking at two lines of the musician's opening
exposition:

> Now, if I remember correctly, there's a song that used to open the portal to
> the Ink Well, where songwriters found the ink to write all the songs of the
> world.

This is 160 characters × 0.02 seconds = 3.2 seconds.

> The songs can make wonderful things happen... make items or creatures appear,
> or even open portals to other regions of Threadbare.

This is 130 characters with an ellipsis, which is treated as 3 full stops, i.e.
3 pauses. 130 * 0.02 + 3 * 0.03 = 3.5 seconds.

If we adjust the two parameters to be twice as fast, these become 1.6 seconds
and 1.75 seconds, respectively. This to me makes the wordier dialogue snappier
without feeling rushed.