ci: Use treeless clone for export workflow
When building & exporting the game, we need the full history of commits so that
`git describe` works, but we do not need the contents of any historical commits.
It is wasteful to download it all, only to discard it as soon as the clone is
finished.

Use a treeless clone, which fetches the full commit history, but only downloads
trees (and blobs, i.e. files) when needed.

https://github.blog/open-source/git/get-up-to-speed-with-partial-clone-and-shallow-clone/