Add ThrowProjectileBehavior
A behavior node to throw (spawn or instantiate) projectile scenes. The
projectile scenes can be any Node2D, but if it is a RigidBody2D then an
initial impulse will be applied.

The projectile scene is picked randomly from the exported array. It
could be added at the top of everything else in the current scene (ie.
in the scene root) or as child of a reference node.

This is intended as an easy way to make a character throw projectiles.