Grappling hook: Add Void shaping the tiles in start sublevel
Make it split the player from the Spirit thread that is so close,
shaping the tiles.

Use hidden blocking tiles in a separate layer, so the player can't
actually move to unwanted places. This could be improved in the future
by being a cut-scene or triggered through dialogue, as during dialogue
the player can't walk.

Helps https://github.com/endlessm/threadbare/issues/1376