Grappling hook: Add Void going back and forth
In powerup sublevel. Add it to the powerup sublevel, going back and
forth the big gap that can only be crossed after adquiring the longer
thread. Make it patrol slowly, so this is the first introduction of Void
enemies that have to be dodged by timing the jump.

Helps https://github.com/endlessm/threadbare/issues/1376