git_lfs_checker: Look up icon in config
This plugin works by opening a file that we know is stored with Git LFS
and checking if it looks like a Git LFS pointer, which indicates that
Git LFS is not set up correctly. The file in question was previously the
project icon, identified by a UID hardcoded in the plugin.

We are going to replace the icon, and unless special care is taken, the
new icon will not have the same UID.

Rather than hardcoding the uid:// path to a file to check, get the path
from the project settings.