Replace icon
For the game itself, this is quite easy: update the configuration to
reference a different icon.

For the Flatpak build this is actually a little fiddly because the old
filename was hardcoded in several places, including in the other version
of the Flatpak manifest that lives on Flathub. So, introduce a new
'install-metadata' script in the linux/ directory, which operates only
on files in that directory, resolving the icon's symlink if it is one.
Invoke this from the manifest in this repo. Include the script in the
tarball that is published with releases and consumed by the Flathub
version of the manifest (which we'll update later). Dereference symbolic
links when creating that tarball.

This should mean that the next time we change the icon, we only have to
update 2 things:

1. project.godot
2. The symbolic link at linux/icon-256.png

Helps https://github.com/endlessm/threadbare/issues/1338