Grappling hook: Allow going back from 2nd sublevel
Add a teleport and corresponding spawn point.

Also adjust the original teleport and tiles in the "start" sublevel so
the player doesn't look like is walking in front of a wall while the
transition runs. And also make the exit transition consistent in this
teleport, making it wipe left to right.

https://github.com/endlessm/threadbare/issues/1379