Grappling hook: Play victory sound on door open
In the third sublevel, there is a door that opens. The door component
has an export to play a victory fanfare sound. Adjust the script too,
because this sound doesn't play using the opened setter, but using the
open() function.

Also commit a change in the door scene updating the sound assets paths.