Void spreading enemy: Add prop that takes time to ingest
Add a pile of books that has a float metadata "ingest_time" and is set
to 2.0. This will be used to stop the Void spreading enemy for that
amount of seconds.

For testing, this prop is also hookable and will move towards the
player, colliding with it. So the player could change its place or bring
it from other "islands" with the grappling hook.