Void spreading enemy: Add ingesting state
In TileMapCover: Report the nodes as they are being consumed. And fix
the function documentation comment, previously it was also returning
true or false if nodes (not tiles) were consumed.

In the enemy new ingesting state, it does nothing more than waiting.
Transition to this state if the metadata of a consumed node says to be
ingested, and set the waiting to that amount of time.

Also emit particles when not the enemy is not moving. The previous
solution for emitting particles when not consuming stuff was
distance-based. Now if the enemy is standing in the new INGESTING state,
nothing would be displayed.