Void runner: Allow enemy to consume starting starting island
Previously the enemy, once triggered, was not able to follow you back
onto the starting islands. So if you triggered the enemy then went left
back to the starting island, you would be stuck: the enemy would not
follow (and eat) you, and there's nothing else you can do.

Expand the navigation polygon to include those two islands. Adjust the
Void layer to add the entities on those islands to the nodes that can be
consumed. Unnest the forest from the Node2D that groups the other props
on the island to make this work.

Resolves https://github.com/endlessm/threadbare/issues/1398