Fill-matching game: Factor out dialogue and setting player mode
Except for turning the player to cozy (interact) mode at the end.
Because although this can be done in each level scene by listening to
the goal_reached signal, it is not worth doing it now as it's soon going
to be revisited. So add a TODO note instead.

By using cinematic and (almost) not affecting the player, this is now
hopefully more exemplar for other kinds of game logic involving barrels,
throwing enemies and projectiles.

FillGameLogic: Remove intro dialogue, allow to autostart.

Ink combat level 1: Add cinematic node. Connect the finished signal of
the cinematic to the start of the logic. Set player to repel from the
start.

Ink combat levels 2 and 3: Set player to repel from the start.
Automatically start the game logic.

NO_EDIT combat, Stella combat, El juguete perdido combat: Do as ink
combat level 1.

Also in Stella, set the player name to "Stella".

Do nothing for shjourney, it has its own scripts that started as copies
of FillGameLogic.