Grappling hook level: Add intro cinematic
The monk tells the StoryWeaver where to go, but suddenly the Void erases
the path. The monk gives her grappling hook (tool name is placeholder)
to the StoryWeaver so they can navigate in the place.

Note: The start and end grappling hook levels will have to be redesigned
because the thread is below the dialogue.

Helps https://github.com/endlessm/threadbare/issues/1410