Stella combat: Reinstate stone stairs
Previously, the two water layers had a border of stair tiles underneath
the rocks at the boundary of the level. This looks weird because we have
changed the stairs to be wooden. It is also surprising to have these in
the water layers (particularly twice).

Split them to their own layer, placed just once. Define a tileset for
this layer which uses the Tiny Swords elevation tiles, which contains
the stone stairs that were used before.