GameState: add current_quest property
Previously, only the path to the current Quest resource was stored in
GameState, and that only in the ConfigFile instance.

Add a `current_quest` property to the GameState. Update this in each
place where the path to the current quest is changed in the ConfigFile.
Restore the Quest instance when restoring state.

This will be useful for other bits of code to refer to the current
quest.