Move 'threads incorporated' dialogue from eternal_loom to elders
When the player returns to Fray's End after completing the Sokoban
puzzles, a brief piece of dialogue is shown where the Elders congratule
the player.

Previously in this dialogue the speaker was “Elders”. I don't think the
elders are a hive mind so it's weird to attribute speech to both of them
at once.

Move the dialogue to each elder's dialogue file, with the specific elder
named as the speaker. Add a congratulate_player() method on the Elder
class to trigger it.

Have the Eternal Loom look up the correct elder for the current quest,
and call congratulate_player() on it, rather than playing the dialogue
directly.