storyquest_bootstrap: Copy TileSets
Previously, the StoryQuest bootstrapping tool did not duplicate the TileSet(s)
used by the NO_EDIT template levels, because the TileSet resource is outside the
NO_EDIT directory. It's very tempting to just edit the TileSet that a layer uses
in-place, without stopping to duplicate it (or make a new TileSet). This is a
common reason why past StoryQuest submissions have needed revision before being
accepted.

When traversing the NO_EDIT template, consider any TileSet resource as something
to be duplicated, even if it is not in the NO_EDIT folder. Duplicate any such
tilesets from outside the NO_EDIT folder to a tiles/ subfolder of the new quest,
and prefix their names with "questname_" like other files.

Resolves https://github.com/endlessm/threadbare/issues/1348