Grappling hook level: Redesign start and end scenes
So the magical thread is not occluded by the dialogue. And to give it
more narrative sense:

- From the Void runner level, there is an ellipsis.

- The StoryWeaver is about to grab a magical thread and be teleported.

- But the Void ruins their plans. Has to take an alternative path.

- At the end of the level, the StoryWeaver comes from the other
  direction.

Add a spawn point to the end sublevel and use it from the previous
(the powerup) sublevel just to use its "look at side on spawn" property.

Helps https://github.com/endlessm/threadbare/issues/1410