Player: Add defeat() method
Previously the code to mark the player as defeated, wait a few seconds,
then reload the current scene was duplicated between two kinds of enemy.
In addition, it was not idempotent: if a second guard caught the player
during the 2-second pause, the scene would reload twice.

Add a defeat() method to the player that does nothing if the mode is
already DEFEATED, and otherwise carries out the same set of steps. Call
this in the two relevant places.

It's a bit of a kludge to have the `falling` parameter but I think this
is still an improvement.

Resolves https://github.com/endlessm/threadbare/issues/1334