Add editor script to update TileSets in all scenes
This walks all scenes in the project, finding those which depend on
scenes/tileset.tres (the previous, unified TileSet). For each scene, it examines
each TileMapLayer that uses the old TileSet. If it only uses atlases from a
single new, split TileSet, it updates the TileSet. If it uses two or more, or
zero, a warning is logged and the script continues.

https://github.com/endlessm/threadbare/issues/1336