Mothsache enemy: Update Aseprite source file with walking animation (#1462)
The Aseprite file for the walking animation has been updated with the
requested corrections to the action line. This change includes
adjustments to the torso and pelvis for a more natural-looking movement.
They will be able to see in more detail how to make improvements to the
aseprite file and implement it in a scenario if they wish.

Fixes: #1360